Add Carousel to Layout folder + clean code
