+++ libs/ui/packages/react/package.json
@@ -78,7 +78,7 @@
     "chart.js": "^3.5.1",
     "chartjs-adapter-moment": "^1.0.0",
     "color": "^4.0.0",
-    "embla-carousel-react": "8.0.0-rc19",
+    "embla-carousel-react": "8.0.0-rc20",
     "moment": "^2.29.1",
     "react-chartjs-2": "^3.3.0",
     "react-is": "^17.0.2",
