--- libs/ui/packages/react/src/components/carousel/index.tsx
@@ -1,1 +1,0 @@
-export { default as Carousel } from "./carousel";
