+++ libs/ui/packages/react/src/components/index.ts
@@ -1,5 +1,4 @@
 export * from "./asorted";
-export * from "./carousel";
 export { default as Chart } from "./Chart";
 export * from "./cta";
 export * from "./form";
