+++ libs/ui/packages/react/src/components/layout/Carousel/Carousel.stories.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import Carousel from "./carousel";
+import Carousel from "./";
 
 const CarouselStory = () => {
   return (
