+++ libs/ui/packages/react/src/components/layout/Carousel/Footer/Pagination/index.tsx
