+++ libs/ui/packages/react/src/components/layout/Carousel/Footer/Pagination/types.ts
