+++ libs/ui/packages/react/src/components/layout/Carousel/Footer/index.tsx
