+++ libs/ui/packages/react/src/components/layout/index.ts
@@ -4,4 +4,5 @@ export { default as Grid } from "./Grid";
 export { default as Popin } from "./Popin";
 export { default as Side } from "./Side";
 export { default as Drawer } from "./Drawer";
+export { default as Carousel } from "./Carousel";
 export { default as VerticalTimeline } from "./List/VerticalTimeline";
